## Summary

`Core/Libraries/Source/WWVegas/WW3D2/dx8wrapper.cpp` now compiles off
Windows at 64-bit, without
defining `_WIN32` and without a stub. It was the last non-cut compile
failure in the level 1-4 native
build, and because it holds the whole `DX8Wrapper::*` surface, its
failure alone accounted for ~86 of
the 108 `compile-blocked` unresolved symbols.

Measured, `clang++-14` on Ubuntu 22.04, levels 1-4 shimmed with
`--strict-link`:

| levels 1-4, shimmed | before (`6df9b180a`) | after |
|---|---:|---:|
| objects | 968/972 | **969/972** |
| compile failures | 4 | **3** — all three the cut-scope GameSpy units
(slice 2) |
| strict-link unresolved | 250 | **173** |
| `compile-blocked` | 108 | **18** (the GameSpy remainder) |
| `no-definition-anywhere` | 9 | **22** — see below, this is the honest
cost |
| strict link produces an executable | no | **still no** |

173, not the estimated ~165: the estimate assumed the 108 → 18 drop was
the only movement, but 13
symbols that were hiding behind the compile failure became visible at
the same time.

### The complete diagnostic list, which was step one

The `DriverVersion` member DirectX gates on `_WIN32` itself was already
routed through the D3D8
accessor by #73, so it was not the blocker. Compiling the file alone
with `-ferror-limit=0` and the
harness' own flags gave two classes:

1. Win32 window/GDI spellings absent from
`scripts/native-port-shims/windows.h`: `GetWindowLong`,
`GWL_STYLE`, `AdjustWindowRect`, `MONITORINFO`, `MonitorFromWindow`,
`GetMonitorInfo`,
`MONITOR_DEFAULTTOPRIMARY`, `SWP_NOZORDER`, `GetDesktopWindow`,
`SetDeviceGammaRamp` — the
windowed-device sizing path and the gamma fallback. Declared in the
shim, under the Win32
   spelling; no call site rewritten.
2. One real 64-bit defect, in the snapshot validation path:

```diff
-		unsigned long passes=0;      // 32-bit on MSVC, 64-bit at LP64
+		DWORD passes=0;              // what ValidateDevice(DWORD*) actually takes
 		DX8CALL_RAW_HRES(ValidateDevice(&passes),res);
```

Both are behaviour-neutral on Windows: one is a header only the
non-Windows probe reads, the other is
a local spelled as the API spells it (`DWORD` == `unsigned long` on
32-bit MSVC).

### `no-definition-anywhere` grew 9 → 22, and nothing is stubbed

Exactly the #65 / `IsIconic` effect: a compiling translation unit stops
hiding its own calls. The 13
newly *visible* symbols are reported, not implemented, and none belongs
to this slice:

- `AdjustWindowRect`, `GetClientRect`, `GetDesktopWindow`,
`GetMonitorInfoA`, `GetWindowLongA`,
`MonitorFromWindow`, `SetWindowPos`, `SetDeviceGammaRamp` → the
window/input + GDI seam
(`WWLib/platform/platform_win32_user.cpp`), where
`GetCursorPos`/`IsIconic`/`ScreenToClient`/
  `SetCursor` already sit in the same budget.
- `D3DXCreateTexture`, `D3DXCreateCubeTexture`,
`D3DXCreateVolumeTexture`,
`D3DXCreateTextureFromFileExA`, `D3DXGetErrorStringA` → the D3DX slices
(`renderer/methods-and-d3dx`, `renderer/d3dx8-math-entrypoints`), which
already own the four D3DX
  symbols in that pile.

Consequently the level 1-4 baseline's `no-definition-anywhere` ratchet
and
`win32-undefined-budget.json` are widened by exactly those eight Win32
names — deliberately, in the
commit that earned the rest, with the reasoning in
`docs/porting/dx8wrapper-native-compile.md`. If
that trade is not wanted, the alternative is leaving the file uncompiled
and the 86 `DX8Wrapper::*`
definitions lost, which is strictly worse.

### Verification

- `check-native-build-baseline.py` (levels 1-4, strict link),
`check-win32-undefined.py`,
`check-generated-baselines.py`, `check-d3d8-surface.py` (direct-call
budget still 0),
`porting-status.py --check`, `window-input-scan.py --check`,
`check-window-scancodes.py`,
`check-window-seam-wiring.py`, `actionlint`: all green. Levels 1-3 are
unchanged (838/841, 548), so
  that baseline is untouched.
- Renderer spike on lavapipe with the validation layer loaded and
silent: pixel compare 0/480000
pixels differ, `zh-fixedfunc-tests`/`zh-resource-lock-tests` 0 failures
with and without
  `ZH_SPIKE_NO_VIEW_SWIZZLE=1`, `check-backend-coverage.py` exact.
- Wine/VC6 Windows oracle build (`scripts/docker-build.sh`) was run for
this commit because it touches
a live renderer source; the result is reported in a comment on this PR.
- `flake8 scripts/` fails on two pre-existing E501 lines in
`scripts/xfer-blob-audit.py`
(untouched here, from #27, not gated in CI). Reported, not fixed — not
this slice's file.

Ownership row added to `docs/porting/concurrent-slices.md`. Borrowed and
disclosed: the generated
level 1-4 baseline + `native-build-report.md` (regenerated by the
scripts, never hand-edited), the
Win32 budget names, and a five-line "superseded" note at the top of
`docs/porting/startability.md`
(no figure in its body rewritten).



Link to Devin session:
https://app.devin.ai/sessions/c53ef1ea95f748c4917712c258660fbd
Requested by: @willhoff2

+++ Core/Libraries/Source/WWVegas/WW3D2/dx8wrapper.cpp
@@ -2028,7 +2028,7 @@ void DX8Wrapper::Draw(
 
 #ifdef MESH_RENDER_SNAPSHOT_ENABLED
 	if (WW3D::Is_Snapshot_Activated()) {
-		unsigned long passes=0;
+		DWORD passes=0;
 		SNAPSHOT_SAY(("ValidateDevice:"));
 		HRESULT res;
 		DX8CALL_RAW_HRES(ValidateDevice(&passes),res);
